+++ NEWS.md
@@ -1,5 +1,9 @@
 # broom.helpers (development version)
 
+**Fix**
+
+- Fix in `supported_models`
+
 # broom.helpers 1.14.0
 
 **New features**